requirements.txt
# Основные зависимости для быстрой интеграции
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
chromadb==0.4.18
google-generativeai==0.3.2
python-multipart==0.0.6
jinja2==3.1.2
pydantic==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
python-docx==1.1.0
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
aiofiles==23.2.1
python-dotenv==1.0.0